Fix withdrawals tests
