+++ tests/core/pyspec/eth2spec/test/helpers/epoch_processing.py
@@ -28,8 +28,7 @@ def get_process_calls(spec):
             'process_participation_record_updates'
         ),
         'process_sync_committee_updates',  # altair
-        'process_full_withdrawals',  # capella
-        'process_partial_withdrawals',  # capella
+        'process_withdrawals_into_queue',  # capella
         # TODO: add sharding processing functions when spec stabilizes.
     ]
